backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
requests>=2.31.0
aiohttp>=3.9.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
web3==6.15.1
python-dotenv==1.0.1
flwr>=1.10.0
